inspection-app/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
WTForms==3.0.1
Bcrypt==1.0.1
WeasyPrint==58.1
python-dotenv==1.0.0